pyproject.toml
[project]
name = "ai-news-automation"
version = "0.1.0"
description = "Automated pipeline to scrape, rank, test, and distribute AI news and tools"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "requests>=2.32.0",
    "feedparser>=6.0.0",
    "praw>=7.8.0",
    "pyyaml>=6.0.0",
    "rapidfuzz>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
browser = [
    "playwright>=1.48.0",
]

[project.scripts]
ainews = "run_pipeline:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
